Fixing a few complier warnings

+++ Gbox420_Nano_Hempy/Settings.h
@@ -45,11 +45,11 @@ static const uint8_t Version = 5; ///< Increment this when you make a change in 
     
     struct HempyBucketSettings
     {
-      HempyBucketSettings( float StartWeight = 0.0, float StopWeight = 0.0, bool SeedlingMode = false, uint8_t SeedlingWaterFreqency = 0) : StartWeight(StartWeight), StopWeight(StopWeight), SeedlingMode(SeedlingMode), SeedlingWaterFreqency(SeedlingWaterFreqency)   {}
+      HempyBucketSettings( float StartWeight = 0.0, float StopWeight = 0.0, bool SeedlingMode = false, uint16_t SeedlingWaterFreqency = 0) : StartWeight(StartWeight), StopWeight(StopWeight), SeedlingMode(SeedlingMode), SeedlingWaterFreqency(SeedlingWaterFreqency)   {}
       float StartWeight; ///Start watering below this weight
       float StopWeight;  ///Stop watering above this weight
       bool SeedlingMode; ///Enables timed waterings to prevent the soil from drying out with seedlings
-      uint8_t SeedlingWaterFreqency; ///Watering frequency in MINUTES when SeedlinMode is enabled
+      uint16_t SeedlingWaterFreqency; ///Watering frequency in MINUTES when SeedlinMode is enabled
     };
     struct HempyBucketSettings Bucket1 = { .StartWeight = 4.2, .StopWeight = 6.9, .SeedlingMode = false, .SeedlingWaterFreqency= 720};
     struct HempyBucketSettings Bucket2 = { .StartWeight = 4.2, .StopWeight = 6.9, .SeedlingMode = false, .SeedlingWaterFreqency= 720};
